Add Blyth-Still-Casella interval

+++ clopper-pearson.cc
@@ -1,7 +1,9 @@
+#include <stdio.h>
+
 #include "clopper-pearson.hh"
 #include "binary-search.hh"
 
-ClopperPearson::ClopperPearson( const int s_N, const real s_alpha )
+ClopperPearson::ClopperPearson( const unsigned int s_N, const real s_alpha )
   : _N( s_N ),
     _alpha( s_alpha )
 {
@@ -14,13 +16,13 @@ static real sum_integer_binomial_range( const unsigned int N,
   real sum( 0 );
 
   for ( unsigned int i = lower; i <= upper; i++ ) {
-    sum += expl( likeln( N, i, p ) );
+    sum += exp( likeln( N, i, p ) );
   }
 
   return sum;
 }
 
-Interval ClopperPearson::limits( const int k )
+Interval ClopperPearson::limits( const unsigned int k )
 {
   Interval ret( 0, 1 );
 
@@ -30,9 +32,7 @@ Interval ClopperPearson::limits( const int k )
   if ( k != 0 ) {
     ret.lower = param_binary_search( Interval( 0, estimate ),
 				     [&] (const real p)
-				     {
-				       return sum_integer_binomial_range( _N, k, _N, p );
-				     },
+				     { return sum_integer_binomial_range( _N, k, _N, p ); },
 				     _alpha / 2.0,
 				     true );
   }
@@ -41,9 +41,7 @@ Interval ClopperPearson::limits( const int k )
   if ( k != _N ) {
     ret.upper = param_binary_search( Interval( estimate, 1 ),
 				     [&] (const real p)
-				     {
-				       return sum_integer_binomial_range( _N, 0, k, p );
-				     },
+				     { return sum_integer_binomial_range( _N, 0, k, p ); },
 				     _alpha / 2.0,
 				     false );
   }
